Small Changes
Rusher now uses health accurate to what it is in the mod (when infecting someone as a headcrab)
If a sci/barney gets a rifle then they will make the alpha reload sound (syncs better with animation)
If they have a shotgun then the normal reload sound is disabled.
(attempted to) fix headcrabs getting stuck inside zombies.